Dataset of targeted and non-targeted cancer drugs

A dataset containing an extensive collection of anti-cancer drugs, i.e. targeted compounds, chemotherapy regimens etc. The data has been harvested from the Open Targets Platform, the NCI Thesaurus, and ChEMBL. Note that each entry corresponds to a "drug synonym-target-indication association"

oncopharmadb

Format

oncopharmadb - A data frame with 418,620 rows and 49 columns:

drug_name - Primary drug name (upper case, NCI Thesaurus)

nci_concept_display_name - Primary drug name (NCI Thesaurus)

drug_type - type of drug molecule (Antibody, small molecule etc)

drug_action_type - main action elicited by drug (antagonist, inhibitor, stabiliser etc)

molecule_chembl_id - ChEMBL compound identifier

drug_max_phase_indication - maximum clinical phase for drug (given indication)

drug_max_ct_phase - maximum clinical phase for drug (any indication)

target_genename - name/description of drug target

target_symbol - symbol of drug target

target_type - type of drug target (single protein, protein family etc.)

target_ensembl_gene_id - Ensembl gene identifier of drug target

target_entrezgene - Entrez gene identifier of drug target

target_uniprot_id - UniProt identifier of drug target

disease_efo_id - EFO (Experimental Factor Ontology) identifier for drug indication

disease_efo_label - EFO (Experimental Factor Ontology) label for drug indication

cui - UMLS metathesaurus identifier (concept unique identifier) for drug indication

cui_name - UMLS metahesaurus name for drug indication

primary_site - primary tumor site/type (obtained through https://github.com/sigven/oncoPhenoMap)

nci_concept_synonym - drug synonym (NCI, Open Targets Platform)

nci_concept_synonym_all - all drug synonyms (NCI Thesurus, Open Targets Platform)

drug_clinical_source - underlying source for drug entry (DailyMed, clinicaltrials.gov, FDA etc.)

drug_approved_indication - logical indicating if drug has an approved indication

drug_blacbox_warning - logical indicating if drug has blackbox warning

drug_year_first_approval - year drug was first approved

drug_clinical_id - drug clinical identifier (clinicaltrials.gov, DailyMed, FDA etc.)

nci_t - NCI Thesaurus identifier

nci_concept_definition - detailed description of drug mechanism-of-action (NCI Thesaurus)

nci_version - version of NCI used for this version of oncoPharmaDB

chembl_version - version of ChEMBL used for this version of oncoPharmaDB

opentargets_version - version of Open Target Platform used for this version of oncoPharmaDB

comb_regimen_indication - type of drug molecule (Antibody, small molecule etc)

immune_checkpoint_inhibitor - logical indicating if drug is an immune checkpoint inhibitor

topoisomerase_inhibitor - logical indicating if drug is a topoisomerase inhibitor

tubulin_inhibitor - logical indicating if drug is a tubulin inhibitor

kinase_inhibitor - logical indicating if drug is a kinase inhibitor

hdac_inhibitor - logical indicating if drug is a HDAC inhibitor

parp_inhibitor - logical indicating if drug is a PARP inhibitor

bet_inhibitor - logical indicating if drug is a BET inhibitor

ar_antagonist - logical indicating if drug is an androgen receptor antagonist

monoclonal_antibody - logical indicating if drug is an antibody

antimetabolite - logical indicating if drug is an antimetabolite

angiogenesis_inhibitor - logical indicating if drug is an angiogenesis inhibitor

alkylating_agent - logical indicating if drug is an alkylating agent

platinum_compound - logical indicating if drug is a platinum compound

anthracycline - logical indicating if drug is an anthracycline

proteasome_inhibitor - logical indicating if drug is a proteasome inhibitor

hormone_therapy - logical indicating if drug is a hormone therapy

hedgehog_antagonist - logical indicating if drug is a hedgehog pathway antagonist

Source

https://targetvalidation.org/

https://www.ebi.ac.uk/chembl/

https://ncithesaurus.nci.nih.gov/ncitbrowser/